TP2/requirements.txt
aiohttp
Pillow 
